About page fix (#107)
* About page fix and linking

* About page fix and linking
