+++ app/about-page/components/CoreValues.js
