+++ app/about-page/components/FounderQuotes.js
