+++ app/about-page/components/HeroSection.js
