+++ app/about-page/components/TeamSection.js
