+++ app/about-page/page.tsx
@@ -1,5 +1,4 @@
-
-import NavBar from "./components/NavBar"
+import NavBar from "./components/NavBar";
 import HeroSection from "./components/HeroSection";
 import InfoSection from "./components/InfoSection";
 import TeamSection from "./components/TeamSection";
@@ -7,11 +6,10 @@ import Footer from "./components/Footer";
 import Misson from "./components/Misson";
 import CoreValues from "./components/CoreValues";
 import FounderQuotes from "./components/FounderQuotes";
-import  Global  from "./components/Global"
+import Global from "./components/Global";
 export default function Home() {
   return (
     <main className="bg-white text-gray-900">
-      
       <NavBar />
       <HeroSection />
       <InfoSection />
@@ -22,5 +20,5 @@ export default function Home() {
       <Global />
       <Footer />
     </main>
-  )
+  );
 }
